+++ packages/react-ui/src/components/RadioButton/index.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 
 import { Hint, Input, InputContainer, RadioButtonComponent, Span } from '#src/components/RadioButton/style';
+import { keyboardKey } from '#src/components/common/keyboardKey';
 
 type Dimension = 'm' | 's';
 
@@ -9,6 +10,8 @@ export interface RadioButtonProps extends React.InputHTMLAttributes<HTMLInputEle
   checked?: boolean;
   /** Отключение кнопки */
   disabled?: boolean;
+  /** Только для чтения */
+  readOnly?: boolean;
   /** Размер радиокнопки */
   dimension?: Dimension;
   /** Дополнительный текст (подсказка), который выводится нижней строкой */
@@ -18,12 +21,31 @@ export interface RadioButtonProps extends React.InputHTMLAttributes<HTMLInputEle
 }
 
 export const RadioButton = React.forwardRef<HTMLInputElement, RadioButtonProps>(
-  ({ children, disabled, error = false, dimension = 'm', extraText, className, ...props }, ref) => {
+  ({ children, disabled, readOnly, error = false, dimension = 'm', extraText, className, ...props }, ref) => {
+    const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
+      if (readOnly) {
+        const code = keyboardKey.getCode(e);
+        if (code === keyboardKey[' ']) {
+          e.preventDefault();
+        }
+      }
+
+      props.onKeyDown?.(e);
+    };
+
     return (
-      <RadioButtonComponent disabled={disabled} className={className} dimension={dimension}>
+      <RadioButtonComponent disabled={disabled} readOnly={readOnly} className={className} dimension={dimension}>
         <InputContainer dimension={dimension}>
-          <Input ref={ref} type="radio" disabled={disabled} dimension={dimension} {...props} />
-          <Span disabled={disabled} dimension={dimension} error={error} />
+          <Input
+            ref={ref}
+            type="radio"
+            disabled={disabled}
+            readOnly={readOnly}
+            dimension={dimension}
+            {...props}
+            onKeyDown={handleKeyDown}
+          />
+          <Span disabled={disabled || readOnly} dimension={dimension} error={error} />
         </InputContainer>
         {children}
         {extraText && (