+++ packages/react-ui/src/index.ts
@@ -6,6 +6,9 @@ export {
   smallGroupBorderRadius,
   mediumGroupBorderRadius,
   largeGroupBorderRadius,
+  DefaultFontColorName,
+  DefaultBackgroundColorName,
+  MainPrimaryColorName,
 } from './components/themes';
 
 export * from './components/Accordion';
@@ -79,6 +82,7 @@ export * from './components/GroupActionsPane';
 export * from './components/skeleton/animation';
 export * from './components/common/utils/refSetter';
 export * from './components/common/utils/checkOverflow';
+export * from './components/common/utils/cleanUpStoriesProps';
 export * from './components/common/keyboardKey';
 export * from './components/common/hooks/useClickOutside';
 export * from './components/common/hooks/useDebounce';
@@ -89,6 +93,7 @@ export * from './components/TagMenu';
 export * from './components/Range';
 export * from './components/Pill';
 export * from './components/IconPlacement';
+export * from './components/Drawer';
 
 export { FontsVTBGroup } from './components/FontsVTBGroup';
 export { DropdownProvider } from './components/DropdownProvider';