+++ packages/react-ui/CHANGELOG.md
@@ -2,6 +2,62 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [4.16.0](https://github.com/AdmiralDS/react-ui/compare/v4.15.1...v4.16.0) (2023-05-23)
+
+
+### Features
+
+* **TextButton:** add iconStart and iconEnd props instead of icon and displayRight ([e077c99](https://github.com/AdmiralDS/react-ui/commit/e077c99303b282a28b0589b6d3aec4fc368b1c88))
+* **TextButtonMenu:** add icon prop ([ee97c8a](https://github.com/AdmiralDS/react-ui/commit/ee97c8a00f1f27bd65c14d6b81cc053c0baf0306))
+* **TextButtonMenu:** add iconStart prop ([1832ef8](https://github.com/AdmiralDS/react-ui/commit/1832ef8842203a6d50d8b36356cd1b17c7e9cf82))
+* **themes:** add shadow style 'Shadow 12' ([628966c](https://github.com/AdmiralDS/react-ui/commit/628966c5e024a30ab440d7eae1bf9a1cfe05f70e))
+
+
+### Bug Fixes
+
+* **Menu:** fix Menu width ([07cd291](https://github.com/AdmiralDS/react-ui/commit/07cd2915e6f838bb7871c9102b5f564e40552703))
+* **Toggle, RadioButton:** fix readOnly state style ([31c11ec](https://github.com/AdmiralDS/react-ui/commit/31c11ece82e1effc61afd9c39cff19ecc88b55ad))
+
+### [4.15.1](https://github.com/AdmiralDS/react-ui/compare/v4.15.0...v4.15.1) (2023-05-19)
+
+
+### Bug Fixes
+
+* **Select:** fix checkbox color in Select drop container for multiple mode ([eaf0b28](https://github.com/AdmiralDS/react-ui/commit/eaf0b28cbad67a6cb135c31fdd09ed4adef3594a))
+
+## [4.15.0](https://github.com/AdmiralDS/react-ui/compare/v4.14.1...v4.15.0) (2023-05-19)
+
+
+### Features
+
+* **Colors:** add colors section in storybook ([bc59e49](https://github.com/AdmiralDS/react-ui/commit/bc59e4912d2e9fcdc8e3e74e1c97ecdd844f6880))
+* **Hint:** change internal vertical paddings ([f30a4fd](https://github.com/AdmiralDS/react-ui/commit/f30a4fd33c31a84ab4c8c41c10fa1a102bc1a479))
+* **Modal:** change internal vertical paddings ([3f44b38](https://github.com/AdmiralDS/react-ui/commit/3f44b3835185b98b6382f8b475e258a355276e5d))
+
+
+### Bug Fixes
+
+* **Drawer:** fix paddings for DrawerTitle and DrawerButtonPanel ([d6af714](https://github.com/AdmiralDS/react-ui/commit/d6af714777836bd6918ebe8250e2816e6b5e71fe))
+* **Select:** fix disabled chip border && remove checkbox redundant readOnly state ([73d8076](https://github.com/AdmiralDS/react-ui/commit/73d8076edd283010ad832598fde2c35527a37b99))
+
+### [4.14.1](https://github.com/AdmiralDS/react-ui/compare/v4.14.0...v4.14.1) (2023-05-15)
+
+## [4.14.0](https://github.com/AdmiralDS/react-ui/compare/v4.13.0...v4.14.0) (2023-05-11)
+
+
+### Features
+
+* **Checkbox, CheckboxField:** add readOnly state ([f530bfe](https://github.com/AdmiralDS/react-ui/commit/f530bfeec2396322712c5e0e9fdb30ffa01db8d4))
+* **Drawer:** add component Drawer and examples for it ([98fe8e6](https://github.com/AdmiralDS/react-ui/commit/98fe8e62a1d389fba73161c1700ad5c057033d62))
+* **RadioButton:** add readOnly state ([cfa47ed](https://github.com/AdmiralDS/react-ui/commit/cfa47ed7913d43d95611003f6f101ba7d7c95058))
+* **Tag,Tags:** add proper kind names ([2115d7c](https://github.com/AdmiralDS/react-ui/commit/2115d7cf3262faabd034a3656bededebbe961842))
+* **Toggle:** add readOnly state ([e69d7da](https://github.com/AdmiralDS/react-ui/commit/e69d7da495deafb17ef7d717a5a07221c9c872f7))
+
+
+### Bug Fixes
+
+* **storybook:** fix example with fonts ([5b5e195](https://github.com/AdmiralDS/react-ui/commit/5b5e1957cac1a2a5c229fa39a2d86be2dbe06900))
+
 ## [4.13.0](https://github.com/AdmiralDS/react-ui/compare/v4.12.0...v4.13.0) (2023-05-03)